Add essential files for new react application
Added essential files for a new React application. These include package.json for the client and server, .gitignore for the client, and .env files to handle environment variables. Also, basic project structure and files for the React app coherently set up to render a rudimentary React application. This includes server and client-side implementations and newly created components which are also imported and used. Added Fib.js for async request handling. In addition, added an index.js file to ensure correct rendering of the React application, and installed necessary dependencies in package.json.